% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycodetfakepkg1.R
\name{tinycodetfakepkg1}
\alias{tinycodetfakepkg1}
\alias{fun_overwritten}
\alias{fun11}
\alias{fun12}
\alias{\%opover\%}
\alias{\%op11\%}
\alias{\%op12\%}
\title{tinycodetfakepkg1}
\usage{
fun_overwritten()

fun11()

fun12()

X \%opover\% A

X \%op11\% A

X \%op12\% A
}
\arguments{
\item{X}{whatever}

\item{A}{whatever}
}
\description{
Bla
}
